Move run record related files into a separate folder.
