+++ console/src/app/modules/app-module/components/runs-table/mongoose-run-status-icon/mongoose-run-status-icon.component.ts
@@ -1,5 +1,5 @@
-import { Component, OnInit, Input, ViewChild, ElementRef } from '@angular/core';
-import { MongooseRunStatus } from 'src/app/core/models/mongoose-run-status';
+import { Component, OnInit, Input, ViewChild, ElementRef } from "@angular/core";
+import { MongooseRunStatus } from "src/app/core/models/mongoose-run-record/mongoose-run-status";
 
 @Component({
   selector: 'app-mongoose-run-status-icon',
